requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.5
django-dotenv==1.4.2
django-environ==0.8.1
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
kombu==5.2.4
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
pyparsing==3.0.9
pytz==2022.1
redis==4.3.3
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
